// src/components/navbar-breadcrumb/enum.ts
enum TagNames {
  DIV = 'div',
  NAV = 'nav',
  A = 'a',
  IMG = 'img',
  SPAN = 'span',
}

enum Styles {
  NAVBAR = 'navbar-container',
  LINK_WRAPPER = 'link-wrapper',
  LINK = 'navbar-link',
  LINKS_CONTAINER = 'links-container',
  IMAGE_WRAPPER = 'home-logo-wrapper',
  IMAGE = 'logo-image',
}

export { TagNames, Styles };